+++ HISTORY.md
@@ -3,6 +3,12 @@
 https://github.com/josdejong/jsoneditor
 
 
+## not yet released, version 4.0.0
+
+- Ace editor and jsonlint are now packed with jsoneditor.js by default.
+- Reworked the source code to CommonJS modules, using `brace` to load Ace.
+
+
 ## 2015-01-25, version 3.2.0
 
 - Implemented shortcut keys `Ctrl+\` to format and `Ctrl+Shift+\` to compact